feat(Cars): 자동차 중복된 이름 생성시 검증하는 로직 추가 및 테스트 작성 (리뷰어님 의견 반영!)
