+++ src/main/java/racingcar/domain/Car.java
@@ -71,12 +71,12 @@ public class Car {
 		if (o == null || getClass() != o.getClass())
 			return false;
 		Car car = (Car)o;
-		return position == car.position && Objects.equals(name, car.name);
+		return Objects.equals(getName(), car.getName());
 	}
 
 	@Override
 	public int hashCode() {
-		return Objects.hash(name, position);
+		return Objects.hash(getName());
 	}
 
 	@Override
